Fix reduce chance to delete drive data

+++ src/core/mount.ts
@@ -2,8 +2,8 @@ import { $ } from "bun";
 import { abort } from "#/utils/abort";
 import { logger } from "#/utils/logger";
 import { spawnInteractive } from "#/utils/spawn";
-import { makeDir, removeDir, removeFile } from "#/utils/path";
 import { configMountDiskTemplate } from "#/core/config/template";
+import { makeDir, removeDir, removeFile, exists } from "#/utils/path";
 
 export type MountDiskOptions = {
   description: string;
@@ -25,12 +25,39 @@ export async function mountDisk(options: MountDiskConfig): Promise<void> {
     await testMount(name, type, uuid);
   }
 
-  const templates = Object.entries(options).map(
-    ([name, config]: [string, MountDiskOptions]): [string, string] => {
-      const fileName = `${parseMountPoint(name).replaceAll("/", "-")}.mount`;
-      const template = configMountDiskTemplate(name, config);
-      return [fileName, template];
-    },
+  // Generate systemd state and file name
+  const templates = await Promise.all(
+    Object.entries(options).map(
+      async ([mountPoint, config]: [string, MountDiskOptions]): Promise<
+        [string, string]
+      > => {
+        const template = configMountDiskTemplate(mountPoint, config);
+        const fileName = `${parseMountPoint(mountPoint).replaceAll("/", "-")}.mount`;
+
+        const isMountPointExist: boolean = await exists(mountPoint);
+        if (isMountPointExist) {
+          try {
+            logger.warn(`Remount ${mountPoint}`);
+            await $`sudo systemctl disable --now ${fileName}`;
+            console.log();
+            await removeFile(`/etc/systemd/system/${fileName}`, true);
+          } catch (e) {
+            console.log()
+            console.error(e)
+            abort(
+              [
+                `Failed to remount ${mountPoint}.`,
+                "The mount may still be in use.",
+                `Run: sudo fuser -vm ${mountPoint}`,
+                "or",
+                `Run: sudo lsof +D ${mountPoint}`,
+              ].join("\n"),
+            );
+          }
+        }
+        return [fileName, template];
+      },
+    ),
   );
 
   const services = await Promise.all(
@@ -43,24 +70,53 @@ export async function mountDisk(options: MountDiskConfig): Promise<void> {
   try {
     await $`sudo systemctl daemon-reload`;
     await $`sudo systemctl enable --now ${services.join(" ")}`;
+    console.log();
   } catch (error) {
     for (const service of services) {
-      await removeFile(`/etc/systemd/system/${service}`, true);
+      const mountPath = `/etc/systemd/system/${service}`;
+      if (await exists(mountPath)) {
+        await removeFile(mountPath, true);
+      }
     }
     abort(`Error enabling systemd units for mount points`);
   }
 }
 
+// unmount drive
+export async function unmountDrive(config: MountDiskConfig) {
+  for (const name of Object.keys(config)) {
+    const service = `${parseMountPoint(name).replaceAll("/", "-")}.mount`;
+    const mountPath = `/etc/systemd/system/${service}`;
+    await $`sudo systemctl disable --now ${service}`;
+    if (await exists(mountPath)) {
+      await removeFile(mountPath, true);
+    }
+  }
+}
+
 async function testMount(mountPoint: string, type: string, uuid: string) {
-  await makeDir(mountPoint, true);
+  let mounted = false;
+  const exist = await exists(mountPoint);
+  if (!exist) await makeDir(mountPoint, true);
+
   try {
-    await spawnInteractive(["sudo", "mount", "-t", type, `/dev/disk/by-uuid/${uuid}`, mountPoint]);
-    await $`sudo umount ${mountPoint}`;
-    await removeDir(mountPoint, true);
+    await spawnInteractive([
+      "sudo",
+      "mount",
+      "-t",
+      type,
+      `/dev/disk/by-uuid/${uuid}`,
+      mountPoint,
+    ]);
+    console.log();
+    mounted = true;
   } catch (error) {
-    await removeDir(mountPoint, true);
-    logger.error(error)
+    console.error(error);
     abort(`Failed to mount ${mountPoint} (${type}, ${uuid})`);
+  } finally {
+    logger.info("Cleanup Tested Drive");
+    if (mounted) await $`sudo umount ${mountPoint}`;
+    if (!exist) await removeDir(mountPoint, true);
   }
 }
